+++ x/epochs/types/hooks.go
@@ -10,9 +10,9 @@ import (
 
 type EpochHooks interface {
 	// the first block whose timestamp is after the duration is counted as the end of the epoch
-	AfterEpochEnd(ctx sdk.Context, epochIdentifier string, epochNumber int64)
+	AfterEpochEnd(ctx sdk.Context, epochIdentifier string, epochNumber int64) error
 	// new epoch is next block of epoch end block
-	BeforeEpochStart(ctx sdk.Context, epochIdentifier string, epochNumber int64)
+	BeforeEpochStart(ctx sdk.Context, epochIdentifier string, epochNumber int64) error
 }
 
 var _ EpochHooks = MultiEpochHooks{}
@@ -25,28 +25,29 @@ func NewMultiEpochHooks(hooks ...EpochHooks) MultiEpochHooks {
 }
 
 // AfterEpochEnd is called when epoch is going to be ended, epochNumber is the number of epoch that is ending.
-func (h MultiEpochHooks) AfterEpochEnd(ctx sdk.Context, epochIdentifier string, epochNumber int64) {
+func (h MultiEpochHooks) AfterEpochEnd(ctx sdk.Context, epochIdentifier string, epochNumber int64) error {
 	for i := range h {
 		panicCatchingEpochHook(ctx, h[i].AfterEpochEnd, epochIdentifier, epochNumber)
 	}
+	return nil
 }
 
 // BeforeEpochStart is called when epoch is going to be started, epochNumber is the number of epoch that is starting.
-func (h MultiEpochHooks) BeforeEpochStart(ctx sdk.Context, epochIdentifier string, epochNumber int64) {
+func (h MultiEpochHooks) BeforeEpochStart(ctx sdk.Context, epochIdentifier string, epochNumber int64) error {
 	for i := range h {
 		panicCatchingEpochHook(ctx, h[i].BeforeEpochStart, epochIdentifier, epochNumber)
 	}
+	return nil
 }
 
 func panicCatchingEpochHook(
 	ctx sdk.Context,
-	hookFn func(ctx sdk.Context, epochIdentifier string, epochNumber int64),
+	hookFn func(ctx sdk.Context, epochIdentifier string, epochNumber int64) error,
 	epochIdentifier string,
 	epochNumber int64,
 ) {
 	wrappedHookFn := func(ctx sdk.Context) error {
-		hookFn(ctx, epochIdentifier, epochNumber)
-		return nil
+		return hookFn(ctx, epochIdentifier, epochNumber)
 	}
 	// TODO: Thread info for which hook this is, may be dependent on larger hook system refactoring
 	err := osmoutils.ApplyFuncIfNoError(ctx, wrappedHookFn)